requirements.txt
pydantic>=2.0
streamlit>=1.30
openai>=2.21.0
python-dotenv>=1.2.0
pytest>=7.0.0
